+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 1.0.0 / 19.12.2017
 ===================
 
-  * (breaking change) Renamed `react-isomorphic-render` to `react-application` since the project outgrew its initial name and it's now more about building a React application rather than just isomorphic React rendering.
+  * (breaking change) Renamed `react-isomorphic-render` to `react-website` since the project outgrew its initial name and it's now more about building a React application rather than just isomorphic React rendering.
   * (breaking change) Dropped old React support, now supports React >= 16 only.
   * (breaking change) Removed `koa` entirely (for simplicity).
   * (breaking change) Added `secure` flag to page rendering service options for HTTPS.